Correct bumpversion search/replace for CHANGELOG
The # in Markdown gets treated as a comment so there is no way to
do multiline search/replace. As a result only a single line with
the version number is replaced. The # Latest headline will have to
be added manually on the next set of changes.

See https://github.com/peritus/bumpversion/issues/199